Add files via upload
The "Online_reservation" Java program is an interactive online reservation system that allows users to log in, book train tickets, and cancel reservations. It uses a HashMap to store login credentials and prompts users to enter their login ID and password. Depending on their choice, users can proceed with booking or cancellation. The program collects necessary details for booking, such as name, train number, class type, date of journey, and station names. For cancellation, users enter their PNR number. Reservation and cancellation functionalities are simulated in the program, and the user is provided with appropriate responses based on their input.